hopefully fixed

+++ assignments/mini-libc/tests/io/stat.c
@@ -14,7 +14,7 @@ static char *itoa(long i, char *b)
 {
 	char const digit[] = "0123456789";
 	char *p = b;
-	int shifter = i;
+	long shifter = i;
 
 	do {
 		++p;
